fixed constructors, added rotation values, both work
